refactor: use form context instead of prop drilling
